backend/requirements.txt
fastapi
uvicorn
python-multipart
deepchecks
modelscan
shap
dice-ml
aequitas
tensorflow-model-analysis
aif360
fairlearn
evidently
google-generativeai
reportlab
pyyaml
scikit-learn==1.3.2
xgboost
pandas
numpy
scipy
pydantic
